+++ packages/cli/src/lib/setup/dbConnection.ts
@@ -32,9 +32,9 @@ export function dbConnect(onlyCheck: boolean, params: Record<string, any>, callb
 /**
  * Connects to the DB or tests the connection.
  *
- * @param onlyCheck
- * @param params
- * @param callback
+ * @param onlyCheck if only connection check should be performed
+ * @param params options used by dbConnect
+ * @param callback called when connection is established or check is done
  */
 export async function dbConnect(
     onlyCheck: boolean | Record<string, any> | DbConnectCallback,
@@ -134,17 +134,16 @@ export async function dbConnect(
                     });
                 } else {
                     console.log(
-                        `No connection to objects ${config.objects.host}:${config.objects.port}[${config.objects.type}]`,
+                        `No connection to objects ${Array.isArray(config.objects.host) ? config.objects.host.join(', ') : config.objects.host}:${Array.isArray(config.objects.port) ? config.objects.port.join(', ') : config.objects.port}[${config.objects.type}]`,
                     );
                     if (onlyCheck) {
-                        callback &&
-                            callback({
-                                objects: objects!,
-                                states: states!,
-                                isOffline: true,
-                                objectsDBType: config.objects.type,
-                                config,
-                            });
+                        callback?.({
+                            objects: objects!,
+                            states: states!,
+                            isOffline: true,
+                            objectsDBType: config.objects.type,
+                            config,
+                        });
                         callback = undefined;
                     } else {
                         return void exitApplicationSave(EXIT_CODES.NO_CONNECTION_TO_OBJ_DB);
@@ -215,7 +214,7 @@ export async function dbConnect(
                         objects = null;
                     }
                     console.log(
-                        `No connection to states ${config.states.host}:${config.states.port}[${config.states.type}]`,
+                        `No connection to states ${Array.isArray(config.states.host) ? config.states.host.join(', ') : config.states.host}:${Array.isArray(config.states.port) ? config.states.port.join(', ') : config.states.port}[${config.states.type}]`,
                     );
                     if (onlyCheck) {
                         callback &&